Update README.md
Line 2: "C++ implementation of a game played with a mappable set of connected rooms" -> "This program is a C++ implementation of a game played with a mappable set of connected rooms". (Explicit)

Line 6: "This is **still under construction**!" -> "This program is not finished". (Explicit and formal)

Line 6-8: "If you would like to help develop the game, check out *CONTRIBUTING.md*. See the file *GameInstructions.txt* for instructions on how to play the game Labyrinth.
-> Check out the *GameInstructions.txt* file for instructions on how to play the game Labyrinth, and see the *CONTRIBUTING.md* if you are interested in contributing to the development of the game.